requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
config==0.5.1
databases==0.5.5
dnspython==2.2.0
email-validator==1.1.3
fastapi==0.74.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.1.0
passlib==1.7.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.5
zipp==3.7.0
